refactor(checker): split destructuring recording helpers
